Cleanup - move BTTokenizationKey into own file & cleanup redundant AuthType enums
